pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "stencil"
version = "0.1.0"
description = "Generate project scaffolding from Jinja2 templates and YAML config"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.scripts]
stencil = "stencil.generate:main"

[tool.setuptools.package-data]
stencil = [
    "config.example.yaml",
    "templates/*",
    "templates/.*",
    "files/.vscode/*",
    "scripts/*",
]
